Responding to Feedback Before Merge
Better organized database concerns. Refactored the index.

Prevent crashing by validating config values on init.

Emptied the prompts file.

Account for non-2XX response when embedding.

Add back a timeout on a noCancelCtx for a background DB call.
