+++ server/modules/assistant/memory.go
@@ -17,14 +17,13 @@ import (
 	"time"
 
 	"github.com/google/uuid"
-	"github.com/security-onion-solutions/securityonion-soc/db"
 	"github.com/security-onion-solutions/securityonion-soc/model"
 	"github.com/security-onion-solutions/securityonion-soc/server"
+	"github.com/security-onion-solutions/securityonion-soc/server/modules/assistant/database"
 	"github.com/security-onion-solutions/securityonion-soc/util"
 	"github.com/security-onion-solutions/securityonion-soc/web"
 
 	"github.com/apex/log"
-	"github.com/pgvector/pgvector-go"
 )
 
 // setupMemoryAgents defines the internal roles backing the background memory
@@ -94,7 +93,7 @@ func (ac *AssistantCoordinator) memoryWorker(ctx context.Context) {
 
 		logger := log.FromContext(ctx).WithField("memoryScanId", uuid.NewString())
 
-		if ac.srv.DB == nil {
+		if ac.store == nil {
 			logger.Warn("no database connection; skipping scan")
 			continue
 		}
@@ -406,7 +405,7 @@ func (ac *AssistantCoordinator) extractFacts(ctx context.Context, details *model
 }
 
 func (ac *AssistantCoordinator) reconcileMemories(ctx context.Context, mems []*model.Memory) (ops []*model.ReconciledMemory, err error) {
-	if ac.srv.DB == nil {
+	if ac.store == nil {
 		return nil, ErrNoDatabase
 	}
 
@@ -440,14 +439,14 @@ func (ac *AssistantCoordinator) reconcileMemories(ctx context.Context, mems []*m
 		var nearby []*model.NearbyMemory
 
 		if canWriteGlobal && ac.maxGlobalMemoriesToReconcile > 0 {
-			nearby, err = ac.FindNearbyMemories(ctx, mem.Embedding, mem.ModelID, GlobalMemories(), WithMinSimilarity(ac.mem2memProximityThreshold), WithLimit(ac.maxGlobalMemoriesToReconcile))
+			nearby, err = ac.store.FindNearbyMemories(ctx, mem.Embedding, mem.ModelID, database.GlobalMemories(), database.WithMinSimilarity(ac.mem2memProximityThreshold), database.WithLimit(ac.maxGlobalMemoriesToReconcile))
 			if err != nil {
 				return nil, err
 			}
 		}
 
 		if mem.TargetUserId != nil && ac.maxUserMemoriesToReconcile > 0 {
-			userNearby, err := ac.FindNearbyMemories(ctx, mem.Embedding, mem.ModelID, UserMemories(*mem.TargetUserId), WithMinSimilarity(ac.mem2memProximityThreshold), WithLimit(ac.maxUserMemoriesToReconcile))
+			userNearby, err := ac.store.FindNearbyMemories(ctx, mem.Embedding, mem.ModelID, database.UserMemories(*mem.TargetUserId), database.WithMinSimilarity(ac.mem2memProximityThreshold), database.WithLimit(ac.maxUserMemoriesToReconcile))
 			if err != nil {
 				return nil, err
 			}
@@ -611,202 +610,6 @@ func (ac *AssistantCoordinator) reconcileMemories(ctx context.Context, mems []*m
 	return ops, nil
 }
 
-type memoryScopeKind int
-
-const (
-	scopeNone memoryScopeKind = iota // zero value: no scope specified — fail closed
-	scopeUser
-	scopeGlobal
-	scopeAllUsers
-)
-
-// MemoryScope determines whose memories FindNearbyMemories searches. Build one
-// with UserMemories, GlobalMemories, or AllMemories; the zero value is invalid
-// and matches nothing. Each scope returns only its exact kind of memory —
-// callers wanting both a user's memories and global memories make two calls.
-type MemoryScope struct {
-	kind         memoryScopeKind
-	targetUserId string
-}
-
-// UserMemories scopes the search to memories targeting the given user.
-func UserMemories(userId string) MemoryScope {
-	return MemoryScope{kind: scopeUser, targetUserId: userId}
-}
-
-// GlobalMemories scopes the search to memories with no target user.
-func GlobalMemories() MemoryScope {
-	return MemoryScope{kind: scopeGlobal}
-}
-
-// AllMemories searches every memory regardless of target user, including other
-// users' memories; reserved for superusers.
-func AllMemories() MemoryScope {
-	return MemoryScope{kind: scopeAllUsers}
-}
-
-type findNearbyMemoriesOpts struct {
-	minSimilarity *float64 // nil = no similarity filter
-	limit         *int     // nil = unlimited
-}
-
-type FindNearbyMemoriesOpt func(*findNearbyMemoriesOpts)
-
-// WithMinSimilarity only returns memories at least this similar to the query
-// embedding.
-func WithMinSimilarity(threshold float64) FindNearbyMemoriesOpt {
-	return func(opts *findNearbyMemoriesOpts) {
-		opts.minSimilarity = &threshold
-	}
-}
-
-// WithLimit caps the number of memories returned.
-func WithLimit(limit int) FindNearbyMemoriesOpt {
-	return func(opts *findNearbyMemoriesOpts) {
-		opts.limit = &limit
-	}
-}
-
-func (ac *AssistantCoordinator) FindNearbyMemories(ctx context.Context, embedding []float32, modelID string, scope MemoryScope, opts ...FindNearbyMemoriesOpt) ([]*model.NearbyMemory, error) {
-	if ac.srv.DB == nil {
-		return nil, ErrNoDatabase
-	}
-
-	options := &findNearbyMemoriesOpts{}
-	for _, opt := range opts {
-		opt(options)
-	}
-
-	args := []any{pgvector.NewVector(embedding), modelID}
-
-	stmt := `SELECT id, created_at, updated_at, last_used_at, user_id, memory_text, session_id, embedding, model_id, target_user_id, 1 - (embedding <=> $1) AS similarity, user_defined, usage_count
-		FROM memories
-		WHERE model_id = $2`
-
-	switch scope.kind {
-	case scopeUser:
-		args = append(args, scope.targetUserId)
-		stmt += fmt.Sprintf(` AND target_user_id = $%d`, len(args))
-	case scopeGlobal:
-		stmt += ` AND target_user_id IS NULL`
-	case scopeAllUsers:
-		// no target clause
-	default:
-		return nil, ErrInvalidMemoryScope
-	}
-
-	if options.minSimilarity != nil {
-		args = append(args, *options.minSimilarity)
-		stmt += fmt.Sprintf(` AND 1 - (embedding <=> $1) >= $%d`, len(args))
-	}
-
-	stmt += ` ORDER BY embedding <=> $1`
-
-	if options.limit != nil {
-		args = append(args, *options.limit)
-		stmt += fmt.Sprintf(` LIMIT $%d`, len(args))
-	}
-
-	var rows db.Rows
-	var err error
-
-	rows, err = ac.srv.DB.Query(ctx, stmt, args...)
-	if err != nil {
-		return nil, err
-	}
-
-	defer rows.Close()
-
-	mems := []*model.NearbyMemory{}
-
-	for rows.Next() {
-		mem := &model.Memory{
-			Auditable: model.Auditable{
-				Kind: "memory",
-			},
-		}
-
-		var sessionId *string
-		var vec pgvector.Vector
-		var similarity float64
-
-		err = rows.Scan(&mem.Id, &mem.CreateTime, &mem.UpdateTime, &mem.LastUsedAt, &mem.UserId, &mem.MemoryText, &sessionId, &vec, &mem.ModelID, &mem.TargetUserId, &similarity, &mem.UserDefined, &mem.UsageCount)
-		if err != nil {
-			return nil, err
-		}
-
-		if sessionId != nil {
-			mem.SessionId = *sessionId
-		}
-
-		mem.Embedding = vec.Slice()
-
-		mems = append(mems, &model.NearbyMemory{
-			Similarity: similarity,
-			Memory:     mem,
-		})
-	}
-
-	if err := rows.Err(); err != nil {
-		return nil, err
-	}
-
-	return mems, nil
-}
-
-func (ac *AssistantCoordinator) AddMemory(ctx context.Context, mem *model.Memory) error {
-	if ac.srv.DB == nil {
-		return ErrNoDatabase
-	}
-
-	var sessionId *string
-	if mem.SessionId != "" {
-		sessionId = &mem.SessionId
-	}
-
-	return ac.srv.DB.QueryRow(ctx, `
-		INSERT INTO memories (user_id, memory_text, session_id, embedding, model_id, target_user_id, user_defined)
-		VALUES ($1, $2, $3, $4, $5, $6, $7)
-		RETURNING id, created_at, updated_at`,
-		mem.UserId, mem.MemoryText, sessionId, pgvector.NewVector(mem.Embedding), mem.ModelID, mem.TargetUserId, mem.UserDefined).
-		Scan(&mem.Id, &mem.CreateTime, &mem.UpdateTime)
-}
-
-func (ac *AssistantCoordinator) UpdateMemory(ctx context.Context, mem *model.Memory) error {
-	if ac.srv.DB == nil {
-		return ErrNoDatabase
-	}
-
-	if mem.Id == "" {
-		return fmt.Errorf("cannot update memory without an id")
-	}
-
-	var sessionId *string
-	if mem.SessionId != "" {
-		sessionId = &mem.SessionId
-	}
-
-	return ac.srv.DB.QueryRow(ctx, `
-		UPDATE memories
-		SET memory_text = $2, session_id = $3, embedding = $4, model_id = $5, target_user_id = $6, updated_at = now()
-		WHERE id = $1
-		RETURNING updated_at, last_used_at, usage_count`,
-		mem.Id, mem.MemoryText, sessionId, pgvector.NewVector(mem.Embedding), mem.ModelID, mem.TargetUserId).
-		Scan(&mem.UpdateTime, &mem.LastUsedAt, &mem.UsageCount)
-}
-
-func (ac *AssistantCoordinator) DeleteMemory(ctx context.Context, id string) error {
-	if ac.srv.DB == nil {
-		return ErrNoDatabase
-	}
-
-	if id == "" {
-		return fmt.Errorf("cannot delete memory without an id")
-	}
-
-	return ac.srv.DB.Exec(ctx, `DELETE FROM memories WHERE id = $1`, id)
-}
-
 func (ac *AssistantCoordinator) applyMemories(ctx context.Context, memories []*model.ReconciledMemory) (created int, updated int, deleted int, errMap map[string]string) {
 	errMap = map[string]string{}
 
@@ -817,17 +620,17 @@ func (ac *AssistantCoordinator) applyMemories(ctx context.Context, memories []*m
 
 		switch action {
 		case "ADD":
-			err = ac.AddMemory(ctx, mem.Memory)
+			err = ac.store.AddMemory(ctx, mem.Memory)
 			if err == nil {
 				created++
 			}
 		case "UPDATE":
-			err = ac.UpdateMemory(ctx, mem.Memory)
+			err = ac.store.UpdateMemory(ctx, mem.Memory)
 			if err == nil {
 				updated++
 			}
 		case "DELETE":
-			err = ac.DeleteMemory(ctx, mem.Memory.Id)
+			err = ac.store.DeleteMemory(ctx, mem.Memory.Id)
 			if err == nil {
 				deleted++
 			}
@@ -883,7 +686,7 @@ func buildMemoryExtractTranscript(details *model.AssistantSessionDetails) string
 }
 
 func (ac *AssistantCoordinator) fetchMemoriesForPrompt(ctx context.Context, content string) (user []*model.NearbyMemory, global []*model.NearbyMemory, err error) {
-	if ac.srv.DB == nil {
+	if ac.store == nil {
 		return nil, nil, ErrNoDatabase
 	}
 
@@ -922,30 +725,18 @@ func (ac *AssistantCoordinator) fetchMemoriesForPrompt(ctx context.Context, cont
 	modelUsed := resp.Model
 
 	if ac.maxUserMemoriesToInclude > 0 {
-		user, err = ac.FindNearbyMemories(ctx, emb, modelUsed, UserMemories(userId), WithMinSimilarity(ac.mem2msgProximityThreshold), WithLimit(ac.maxUserMemoriesToInclude))
+		user, err = ac.store.FindNearbyMemories(ctx, emb, modelUsed, database.UserMemories(userId), database.WithMinSimilarity(ac.mem2msgProximityThreshold), database.WithLimit(ac.maxUserMemoriesToInclude))
 		if err != nil {
 			return nil, nil, err
 		}
 	}
 
 	if canReadGlobal && ac.maxGlobalMemoriesToInclude > 0 {
-		global, err = ac.FindNearbyMemories(ctx, emb, modelUsed, GlobalMemories(), WithMinSimilarity(ac.mem2msgProximityThreshold), WithLimit(ac.maxGlobalMemoriesToInclude))
+		global, err = ac.store.FindNearbyMemories(ctx, emb, modelUsed, database.GlobalMemories(), database.WithMinSimilarity(ac.mem2msgProximityThreshold), database.WithLimit(ac.maxGlobalMemoriesToInclude))
 		if err != nil {
 			return nil, nil, err
 		}
 	}
 
 	return user, global, nil
-}
-
-func (ac *AssistantCoordinator) countMemoryUsage(ctx context.Context, ids []string) error {
-	if ac.srv.DB == nil {
-		return ErrNoDatabase
-	}
-
-	if len(ids) == 0 {
-		return nil
-	}
-
-	return ac.srv.DB.Exec(ctx, `UPDATE memories SET usage_count = usage_count + 1, last_used_at = $2 WHERE id = ANY($1)`, ids, time.Now())
 }
